TaskServiceBase clears actions on last task active waypoint. Following waypoint has all the actions since it is cloned before clearing the actions.
This prevents end of behavior actions from executing until after the last waypoint has been passed. This was an issue particularly with Surface Vehicles. The small turn radius frequently results in straight line plans. AMASE then appears to interpret a loiter action for vehicles as  stopping in place. This results in point search tasks appearing to not do anything.