ImpactLineSearchTask will only move the camera for air vehicles.
For the case of ground and surface vehicles. There is no altitude difference between the path and the vehicle, resulting in commands to point the gimbal at the vehicles location. This more often than not results in the camera pointing the wrong way. Usually behind the vehicle. The default camera facing forward Still sees all the path and is more correct.