Logging Changes. The configuration value "isDataTimestamp" is used to pick if log files should include the timestamp. If this is false, files are overwritten. If this is true, new files are generated.
This was put in the top-level configuration to match the isLoggingThreadID configuration option
Removed "isDataTimestamp" from ServiceBase because it appeared to only be used in the MessageLoggerDataService. That now points to the global config option.
To prevent databases from being appended to, database log files are deleted if they exists.