AutomationValidator updates. Reworked OnResponseTimeout. Took out call to checkTasksInitialized in the sendResponseError method.
OnResponseTimout used to cause issues restarting the timer because it executed on the same thread. This would cause a second response to hang if the first timed out. This is solved by letting the ZMQ processing thread restart the timer.

The call to checkTaskInitialized was taken out because it was not needed.