[Validator] Fixing missing estimates for all but the first task on ImpactAutomationResponses.
All that had to be done was set the previous task ID. This was not possible with only looking at the automation request task list.

Chose to use mission commands because it could allow removing the saved request message which is not in the open source stuff. This does feel verbose. Especially since the BatchSummaryService already iterates over the waypoints in the same way.